refactor: 각 객체 성격에 따른 패키지 분리
