+++ src/main/kotlin/racingCar/Application.kt
@@ -1,5 +1,10 @@
 package racingCar
 
+import racingCar.controller.RaceGame
+import racingCar.model.Referee
+import racingCar.view.InputView
+import racingCar.view.OutputView
+
 fun main() {
     val inputView = InputView()
     val outputView = OutputView()
