+++ src/main/kotlin/racingCar/model/Car.kt
@@ -1,4 +1,4 @@
-package racingCar
+package racingCar.model
 
 class Car(
     val name: String,
