+++ src/test/kotlin/racingCar/CarTest.kt
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingCar.model.Car
 
 class CarTest {
     private val car = Car("뭉치")
