+++ src/test/kotlin/racingCar/RefereeTest.kt
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingCar.model.Car
+import racingCar.model.Referee
 
 class RefereeTest {
     private val referee = Referee()
